Rename frontend method to createDemoCourse and move it to course service
